Rename API/framework modules and tests
