+++ api/api/controllers/security_controller.py
@@ -13,8 +13,8 @@ from api.configuration import default_security_configuration
 from api.encoder import dumps, prettify
 from api.models.base_model_ import Body
 from api.models.configuration import SecurityConfigurationModel
-from api.models.security import CreateUserModel, UpdateUserModel, RoleModel, PolicyModel, RuleModel
-from api.models.token_response import TokenResponseModel
+from api.models.security_model import CreateUserModel, UpdateUserModel, RoleModel, PolicyModel, RuleModel
+from api.models.security_token_response import TokenResponseModel
 from api.util import remove_nones_to_dict, raise_if_exc, parse_api_param
 from wazuh import security
 from wazuh.core.cluster.control import get_system_nodes
