+++ api/api/models/__init__.py
@@ -4,4 +4,4 @@
 from __future__ import absolute_import
 
 from api.models.basic_info import BasicInfo
-from api.models.token_response import TokenResponseModel
+from api.models.security_token_response import TokenResponseModel
