+++ api/test/integration/test_agent_DELETE_endpoints.tavern.yaml
