+++ api/test/integration/test_cdb_list_endpoints.tavern.yaml
