+++ api/test/integration/test_decoder_endpoints.tavern.yaml
