+++ api/test/integration/test_rbac_black_active_response_endpoints.tavern.yaml
