+++ api/test/integration/test_rbac_black_agent_endpoints.tavern.yaml
