+++ api/test/integration/test_rbac_black_cdb_list_endpoints.tavern.yaml
