+++ api/test/integration/test_rbac_black_decoder_endpoints.tavern.yaml
