+++ api/test/integration/test_rbac_black_rule_endpoints.tavern.yaml
