+++ api/test/integration/test_rbac_white_active_response_endpoints.tavern.yaml
