+++ api/test/integration/test_rbac_white_cdb_list_endpoints.tavern.yaml
