+++ api/test/integration/test_rbac_white_decoder_endpoints.tavern.yaml
