+++ api/test/integration/test_rbac_white_rule_endpoints.tavern.yaml
