+++ api/test/integration/test_rule_endpoints.tavern.yaml
